requirements.txt
alyahmor==0.2
Arabic-Stopwords==0.4.3
arramooz-pysqlite==0.4.1
asmai==0.1
certifi==2022.12.7
charset-normalizer==3.1.0
codernitydb3==0.6.0
colorama==0.4.6
farasapy==0.0.14
future==0.18.3
idna==3.4
libqutrub==1.2.4.1
maskouk-pysqlite==0.1
mishkal==0.4.1
mysam-tagmanager==0.3.3
Naftawayh==0.4
packaging==23.0
pickleDB==0.9.2
Pillow==9.4.0
PyArabic==0.6.15
pytesseract==0.3.10
qalsadi==0.4.6
requests==2.28.2
six==1.16.0
sylajone==0.3.1
Tashaphyne==0.3.6
tqdm==4.65.0
urllib3==1.26.15
